+++ x/btcstaking/keeper/btc_delegations.go
@@ -48,20 +48,29 @@ func (k Keeper) AddBTCDelegation(ctx sdk.Context, btcDel *types.BTCDelegation) e
 	// save this BTC delegation
 	k.setBTCDelegation(ctx, btcDel)
 
-	// notify subscriber
-	event := &types.EventBTCDelegationStateUpdate{
-		StakingTxHash: stakingTxHash.String(),
-		NewState:      types.BTCDelegationStatus_PENDING,
-	}
-	if err := ctx.EventManager().EmitTypedEvent(event); err != nil {
-		panic(fmt.Errorf("failed to emit EventBTCDelegationStateUpdate for the new pending BTC delegation: %w", err))
+	if err := ctx.EventManager().EmitTypedEvents(types.NewBtcDelCreationEvent(
+		stakingTxHash.String(),
+		btcDel,
+	)); err != nil {
+		panic(fmt.Errorf("failed to emit events for the new pending BTC delegation: %w", err))
 	}
 
 	// NOTE: we don't need to record events for pending BTC delegations since these
 	// do not affect voting power distribution
 	// NOTE: we only insert unbonded event if the delegation already has inclusion proof
 	if btcDel.HasInclusionProof() {
+		if err := ctx.EventManager().EmitTypedEvent(types.NewInclusionProofEvent(
+			stakingTxHash.String(),
+			btcDel.StartHeight,
+			btcDel.EndHeight,
+			types.BTCDelegationStatus_PENDING,
+		)); err != nil {
+			panic(fmt.Errorf("failed to emit EventBTCDelegationInclusionProofReceived for the new pending BTC delegation: %w", err))
+		}
+
 		// record event that the BTC delegation will become unbonded at endHeight-w
+		// This event will be generated to subscribers as block event, when the
+		// btc light client block height will reach btcDel.EndHeight-wValue
 		unbondedEvent := types.NewEventPowerDistUpdateWithBTCDel(&types.EventBTCDelegationStateUpdate{
 			StakingTxHash: stakingTxHash.String(),
 			NewState:      types.BTCDelegationStatus_UNBONDED,
@@ -85,6 +94,7 @@ func (k Keeper) addCovenantSigsToBTCDelegation(
 	parsedUnbondingSlashingAdaptorSignatures []asig.AdaptorSignature,
 	params *types.Params,
 ) {
+
 	// All is fine add received signatures to the BTC delegation and BtcUndelegation
 	btcDel.AddCovenantSigs(
 		covPK,
@@ -95,37 +105,46 @@ func (k Keeper) addCovenantSigsToBTCDelegation(
 
 	k.setBTCDelegation(ctx, btcDel)
 
+	if err := ctx.EventManager().EmitTypedEvent(types.NewCovenantSignatureReceivedEvent(
+		btcDel,
+		covPK,
+		unbondingTxSig,
+	)); err != nil {
+		panic(fmt.Errorf("failed to emit EventCovenantSignatureRecevied for the new active BTC delegation: %w", err))
+	}
+
 	// If reaching the covenant quorum after this msg, the BTC delegation becomes
 	// active. Then, record and emit this event
 	if btcDel.HasCovenantQuorums(params.CovenantQuorum) {
 		if btcDel.HasInclusionProof() {
-			// this BTC delegation does not go through pre-approval flow
-			// notify subscriber that the BTC delegation becomes active
-			event := &types.EventBTCDelegationStateUpdate{
-				StakingTxHash: btcDel.MustGetStakingTxHash().String(),
-				NewState:      types.BTCDelegationStatus_ACTIVE,
-			}
-			if err := ctx.EventManager().EmitTypedEvent(event); err != nil {
-				panic(fmt.Errorf("failed to emit EventBTCDelegationStateUpdate for the new active BTC delegation: %w", err))
+			quorumReachedEvent := types.NewCovenantQuorumReachedEvent(
+				btcDel,
+				types.BTCDelegationStatus_ACTIVE,
+			)
+			if err := ctx.EventManager().EmitTypedEvent(quorumReachedEvent); err != nil {
+				panic(fmt.Errorf("failed to emit emit for the new verified BTC delegation: %w", err))
 			}
 
 			// record event that the BTC delegation becomes active at this height
-			activeEvent := types.NewEventPowerDistUpdateWithBTCDel(event)
+			activeEvent := types.NewEventPowerDistUpdateWithBTCDel(
+				&types.EventBTCDelegationStateUpdate{
+					StakingTxHash: btcDel.MustGetStakingTxHash().String(),
+					NewState:      types.BTCDelegationStatus_ACTIVE,
+				},
+			)
 			btcTip := k.btclcKeeper.GetTipInfo(ctx)
 			k.addPowerDistUpdateEvent(ctx, btcTip.Height, activeEvent)
 		} else {
-			// this BTC delegation goes through pre-approval flow
-			// notify subscriber that the BTC delegation becomes verified
-			event := &types.EventBTCDelegationStateUpdate{
-				StakingTxHash: btcDel.MustGetStakingTxHash().String(),
-				NewState:      types.BTCDelegationStatus_VERIFIED,
+			quorumReachedEvent := types.NewCovenantQuorumReachedEvent(
+				btcDel,
+				types.BTCDelegationStatus_VERIFIED,
+			)
+
+			if err := ctx.EventManager().EmitTypedEvent(quorumReachedEvent); err != nil {
+				panic(fmt.Errorf("failed to emit emit for the new verified BTC delegation: %w", err))
 			}
-			if err := ctx.EventManager().EmitTypedEvent(event); err != nil {
-				panic(fmt.Errorf("failed to emit EventBTCDelegationStateUpdate for the new active BTC delegation: %w", err))
-			}
-			// NOTE: no need to record power dist update event for verified BTC delegations
-			// since it does not have voting power
 		}
+
 	}
 }
 
@@ -139,20 +158,14 @@ func (k Keeper) btcUndelegate(
 	btcDel.BtcUndelegation.DelegatorUnbondingSig = unbondingTxSig
 	k.setBTCDelegation(ctx, btcDel)
 
-	// notify subscriber about this unbonded BTC delegation
-	event := &types.EventBTCDelegationStateUpdate{
-		StakingTxHash: btcDel.MustGetStakingTxHash().String(),
-		NewState:      types.BTCDelegationStatus_UNBONDED,
-	}
-
 	if !btcDel.HasInclusionProof() {
 		return
 	}
 
-	// NOTE: do not generate voting power change event if the delegation
-	// does not have inclusion proof
-	if err := ctx.EventManager().EmitTypedEvent(event); err != nil {
-		panic(fmt.Errorf("failed to emit EventBTCDelegationStateUpdate for the new unbonded BTC delegation: %w", err))
+	// notify subscriber about this unbonded BTC delegation
+	event := &types.EventBTCDelegationStateUpdate{
+		StakingTxHash: btcDel.MustGetStakingTxHash().String(),
+		NewState:      types.BTCDelegationStatus_UNBONDED,
 	}
 
 	// record event that the BTC delegation becomes unbonded at this height