+++ x/btcstaking/types/events.go
@@ -1,6 +1,8 @@
 package types
 
 import (
+	"encoding/hex"
+
 	bbn "github.com/babylonlabs-io/babylon/types"
 )
 
@@ -39,5 +41,76 @@ func NewEventPowerDistUpdateWithUnjailedFP(fpBTCPK *bbn.BIP340PubKey) *EventPowe
 				Pk: fpBTCPK,
 			},
 		},
+	}
+}
+
+func NewInclusionProofEvent(
+	stakingTxHash string,
+	startHeight uint64,
+	endHeight uint64,
+	state BTCDelegationStatus,
+) *EventBTCDelegationInclusionProofReceived {
+	return &EventBTCDelegationInclusionProofReceived{
+		StakingTxHash: stakingTxHash,
+		StartHeight:   startHeight,
+		EndHeight:     endHeight,
+		State:         state,
+	}
+}
+
+func NewBtcDelCreationEvent(
+	stakingTxHash string,
+	btcDel *BTCDelegation,
+) *EventBTCDelegationCreated {
+	return &EventBTCDelegationCreated{
+		StakingTxHash:             stakingTxHash,
+		ParamsVersion:             btcDel.ParamsVersion,
+		FinalityProviderBtcPksHex: btcDel.FinalityProviderKeys(),
+		StakerBtcPkHex:            btcDel.BtcPk.MarshalHex(),
+		StakingTime:               btcDel.StakingTime,
+		StakingAmount:             btcDel.TotalSat,
+		UnbondingTime:             btcDel.UnbondingTime,
+		UnbondingTx:               hex.EncodeToString(btcDel.BtcUndelegation.UnbondingTx),
+		State:                     BTCDelegationStatus_PENDING,
+	}
+}
+
+func NewCovenantSignatureReceivedEvent(
+	btcDel *BTCDelegation,
+	covPK *bbn.BIP340PubKey,
+	unbondingTxSig *bbn.BIP340Signature,
+) *EventCovenantSignatureRecevied {
+	return &EventCovenantSignatureRecevied{
+		StakingTxHash:                 btcDel.MustGetStakingTxHash().String(),
+		CovenantBtcPkHex:              covPK.MarshalHex(),
+		CovenantUnbondingSignatureHex: unbondingTxSig.ToHexStr(),
+	}
+}
+
+func NewCovenantQuorumReachedEvent(
+	btcDel *BTCDelegation,
+	state BTCDelegationStatus,
+) *EventCovenantQuroumReached {
+	return &EventCovenantQuroumReached{
+		StakingTxHash: btcDel.MustGetStakingTxHash().String(),
+		State:         state,
+	}
+}
+
+func NewDelegationUnbondedEarlyEvent(
+	stakingTxHash string,
+) *EventBTCDelgationUnbondedEarly {
+	return &EventBTCDelgationUnbondedEarly{
+		StakingTxHash: stakingTxHash,
+		State:         BTCDelegationStatus_UNBONDED,
+	}
+}
+
+func NewExpiredDelegationEvent(
+	stakingTxHash string,
+) *EventBTCDelegationExpired {
+	return &EventBTCDelegationExpired{
+		StakingTxHash: stakingTxHash,
+		State:         BTCDelegationStatus_UNBONDED,
 	}
 }