+++ CHANGELOG.md
@@ -37,6 +37,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## Unreleased
 
+### Bug Fixes
+
+* [#140](https://github.com/babylonlabs-io/babylon/pull/140) Removed `unbonding`
+and add `verified` to delegation status parse `NewBTCDelegationStatusFromString`.
+
 ### State Machine Breaking
 
 * [#130](https://github.com/babylonlabs-io/babylon/pull/130) Fix bugs in the
@@ -48,7 +53,10 @@ pre-approval flow.
 * [#138](https://github.com/babylonlabs-io/babylon/pull/138) Intercept staking module
 messages inside `authz.MsgExec`.
 
-### Misc Improvements
+### Improvements
+
+* #[131](https://github.com/babylonlabs-io/babylon/pull/131) Add new staking
+events
 
 * [#113](https://github.com/babylonlabs-io/babylon/pull/113) Add multibuild binary
 for upgrade handler `testnet` and `mainnet`.